refactor detector module
