--- Assets/Models/TestImg.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d9c47e79f832e0248a915631840096a8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
