--- Assets/Scripts/AnchorCreatorYoloV3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aee95338b9884b34e8a9040940ad6424
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
