+++ Assets/Scripts/DetectorYolo2.cs
@@ -6,11 +6,9 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
 
-namespace YoloV3
+
+public class DetectorYolo2 : MonoBehaviour, Detector
 {
-    public class DetectorYoloV3 : MonoBehaviour
-{
-    
     public NNModel modelFile;
     public TextAsset labelsFile;
 
@@ -20,48 +18,37 @@ namespace YoloV3
     // ONNX model input and output name. Modify when switching models.
     //These aren't const values because they need to be easily edited on the component before play mode
 
-    //private const string INPUT_NAME = "yolov2-tiny/net1";
-    //private const string OUTPUT_NAME = "yolov2-tiny/convolutional9/BiasAdd";
-    //public string INPUT_NAME = "000_net";
-    //public string OUTPUT_NAME = "023_convolutional";
-
     public string INPUT_NAME;
-    public string OUTPUT_NAME_L;
-    public string OUTPUT_NAME_M;
+    public string OUTPUT_NAME;
 
     //This has to stay a const
-    public const int IMAGE_SIZE = 416;
+    private const int _image_size = 416;
+    public int IMAGE_SIZE { get => _image_size; }
+
 
     // Minimum detection confidence to track a detection
-    public float MINIMUM_CONFIDENCE = 0.25f;
+    public float MINIMUM_CONFIDENCE;
 
     private IWorker worker;
 
-    // public const int ROW_COUNT_L = 13;
-    // public const int COL_COUNT_L = 13;
-    // public const int ROW_COUNT_M = 26;
-    // public const int COL_COUNT_M = 26;
-    public Dictionary<string, int> params_l = new Dictionary<string, int>(){{"ROW_COUNT", 13}, {"COL_COUNT", 13}, {"CELL_WIDTH", 32}, {"CELL_HEIGHT", 32}};
-    public Dictionary<string, int> params_m = new Dictionary<string, int>(){{"ROW_COUNT", 26}, {"COL_COUNT", 26}, {"CELL_WIDTH", 16}, {"CELL_HEIGHT", 16}};
-    public const int BOXES_PER_CELL = 3;
+    public const int ROW_COUNT = 13;
+    public const int COL_COUNT = 13;
+    public const int BOXES_PER_CELL = 5;
     public const int BOX_INFO_FEATURE_COUNT = 5;
 
     //Update this!
     public int CLASS_COUNT;
 
-    // public const float CELL_WIDTH_L = 32;
-    // public const float CELL_HEIGHT_L = 32;
-    // public const float CELL_WIDTH_M = 16;
-    // public const float CELL_HEIGHT_M = 16;
+    public const float CELL_WIDTH = 32;
+    public const float CELL_HEIGHT = 32;
     private string[] labels;
 
     private float[] anchors = new float[]
     {
         // 1.08F, 1.19F, 3.42F, 4.41F, 6.63F, 11.38F, 9.42F, 5.11F, 16.62F, 10.52F  // yolov2-tiny-voc
         //0.57273F, 0.677385F, 1.87446F, 2.06253F, 3.33843F, 5.47434F, 7.88282F, 3.52778F, 9.77052F, 9.16828F // yolov2-tiny
-        //0.57273F, 0.677385F, 1.87446F, 2.06253F, 3.33843F, 5.47434F, 7.88282F, 3.52778F, 9.77052F, 9.16828F  // yolov2-tiny-food
+        0.57273F, 0.677385F, 1.87446F, 2.06253F, 3.33843F, 5.47434F, 7.88282F, 3.52778F, 9.77052F, 9.16828F  // yolov2-tiny-food
         //0.57273F, 0.677385F, 1.87446F, 2.06253F, 3.33843F, 5.47434F, 7.88282F, 3.52778F, 9.77052F, 9.16828F // yolov3
-        10F, 14F,  23F, 27F,  37F, 58F,  81F, 82F,  135F, 169F,  344F, 319F // yolov3-tiny
     };
 
 
@@ -71,12 +58,10 @@ namespace YoloV3
             .Where(s => !String.IsNullOrEmpty(s)).ToArray();
         var model = ModelLoader.Load(this.modelFile);
         // https://docs.unity3d.com/Packages/com.unity.barracuda@1.0/manual/Worker.html
-        //var workerType = WorkerFactory.Type.ComputePrecompiled; // GPU
+        // var workerType = WorkerFactory.Type.ComputePrecompiled; // GPU
         var workerType = WorkerFactory.Type.CSharpBurst;  // CPU
         this.worker = WorkerFactory.CreateWorker(workerType, model);
     }
-
-    
 
 
     public IEnumerator Detect(Color32[] picture, System.Action<IList<BoundingBox>> callback)
@@ -87,13 +72,9 @@ namespace YoloV3
             inputs.Add(INPUT_NAME, tensor);
             yield return StartCoroutine(worker.StartManualSchedule(inputs));
             //worker.Execute(inputs);
-            var output_l = worker.PeekOutput(OUTPUT_NAME_L);
-            var output_m = worker.PeekOutput(OUTPUT_NAME_M);
-            //Debug.Log("Output: " + output);
-            var results_l = ParseOutputs(output_l, MINIMUM_CONFIDENCE, params_l);
-            var results_m = ParseOutputs(output_m, MINIMUM_CONFIDENCE, params_m);
-            var results = results_l.Concat(results_m).ToList();
-
+            var output = worker.PeekOutput(OUTPUT_NAME);
+            Debug.Log("Output: " + output);
+            var results = ParseOutputs(output, MINIMUM_CONFIDENCE);
             var boxes = FilterBoundingBoxes(results, 5, MINIMUM_CONFIDENCE);
             callback(boxes);
         }
@@ -117,20 +98,20 @@ namespace YoloV3
     }
 
 
-    private IList<BoundingBox> ParseOutputs(Tensor yoloModelOutput, float threshold, Dictionary<string, int> parameters)
+    private IList<BoundingBox> ParseOutputs(Tensor yoloModelOutput, float threshold)
     {
         var boxes = new List<BoundingBox>();
 
-        for (int cy = 0; cy < parameters["COL_COUNT"]; cy++)
+        for (int cy = 0; cy < COL_COUNT; cy++)
         {
-            for (int cx = 0; cx < parameters["ROW_COUNT"]; cx++)
+            for (int cx = 0; cx < ROW_COUNT; cx++)
             {
                 for (int box = 0; box < BOXES_PER_CELL; box++)
                 {
                     var channel = (box * (CLASS_COUNT + BOX_INFO_FEATURE_COUNT));
                     var bbd = ExtractBoundingBoxDimensions(yoloModelOutput, cx, cy, channel);
                     float confidence = GetConfidence(yoloModelOutput, cx, cy, channel);
-                    
+
                     if (confidence < threshold)
                     {
                         continue;
@@ -146,7 +127,7 @@ namespace YoloV3
                         continue;
                     }
 
-                    var mappedBoundingBox = MapBoundingBoxToCell(cx, cy, box, bbd, parameters);
+                    var mappedBoundingBox = MapBoundingBoxToCell(cx, cy, box, bbd);
                     boxes.Add(new BoundingBox
                     {
                         Dimensions = new BoundingBoxDimensions
@@ -200,19 +181,19 @@ namespace YoloV3
 
     private float GetConfidence(Tensor modelOutput, int x, int y, int channel)
     {
-        //Debug.Log("ModelOutput " + modelOutput);
+        // Debug.Log("ModelOutput " + modelOutput);
         return Sigmoid(modelOutput[0, x, y, channel + 4]);
     }
 
 
-    private CellDimensions MapBoundingBoxToCell(int x, int y, int box, BoundingBoxDimensions boxDimensions, Dictionary<string, int> parameters)
+    private CellDimensions MapBoundingBoxToCell(int x, int y, int box, BoundingBoxDimensions boxDimensions)
     {
         return new CellDimensions
         {
-            X = ((float)y + Sigmoid(boxDimensions.X)) * parameters["CELL_WIDTH"],
-            Y = ((float)x + Sigmoid(boxDimensions.Y)) * parameters["CELL_HEIGHT"],
-            Width = (float)Math.Exp(boxDimensions.Width) * anchors[6 + box * 2],
-            Height = (float)Math.Exp(boxDimensions.Height) * anchors[6 + box * 2 + 1],
+            X = ((float)y + Sigmoid(boxDimensions.X)) * CELL_WIDTH,
+            Y = ((float)x + Sigmoid(boxDimensions.Y)) * CELL_HEIGHT,
+            Width = (float)Math.Exp(boxDimensions.Width) * CELL_WIDTH * anchors[box * 2],
+            Height = (float)Math.Exp(boxDimensions.Height) * CELL_HEIGHT * anchors[box * 2 + 1],
         };
     }
 
@@ -312,42 +293,4 @@ namespace YoloV3
         }
         return results;
     }
-}
-
-
-public class DimensionsBase
-{
-    public float X { get; set; }
-    public float Y { get; set; }
-    public float Height { get; set; }
-    public float Width { get; set; }
-}
-
-
-public class BoundingBoxDimensions : DimensionsBase { }
-
-class CellDimensions : DimensionsBase { }
-
-
-public class BoundingBox
-{
-    public BoundingBoxDimensions Dimensions { get; set; }
-
-    public string Label { get; set; }
-
-    public float Confidence { get; set; }
-
-    // whether the bounding box already is used to raycast anchors
-    public bool Used { get; set; }
-
-    public Rect Rect
-    {
-        get { return new Rect(Dimensions.X, Dimensions.Y, Dimensions.Width, Dimensions.Height); }
-    }
-
-    public override string ToString()
-    {
-        return $"{Label}:{Confidence}, {Dimensions.X}:{Dimensions.Y} - {Dimensions.Width}:{Dimensions.Height}";
-    }
-}
 }
