--- Assets/Scripts/DetectorYoloV3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3bd27a5c38cf7f24f918e7de7d61d128
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
