--- Assets/Scripts/PhoneARCameraYoloV3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5daca99ad57b8249bb42f798c2b1d42
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
