--- Assets/Scripts/YoloV3Prediction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 65ccbe6ae3b6f3141a9a60b2f79ae7bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
