--- Assets/Scripts/YoloV3Result.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 013337a3e191e2143aa5797caeaacf0a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
